+++ light-client/src/tests.rs
@@ -38,7 +38,6 @@ pub struct Initial {
     pub signed_header: SignedHeader,
     pub next_validator_set: ValidatorSet,
     pub trusting_period: DurationStr,
-    #[serde(with = "tendermint::serializers::time")]
     pub now: Time,
 }
 
@@ -49,7 +48,6 @@ pub struct TestBisection<LB> {
     pub primary: Provider<LB>,
     pub witnesses: Vec<WitnessProvider<LB>>,
     pub height_to_verify: HeightStr,
-    #[serde(with = "tendermint::serializers::time")]
     pub now: Time,
     pub expected_output: Option<String>,
     pub expected_num_of_bisections: usize,