+++ tendermint/src/serializers.rs
@@ -10,4 +10,3 @@ pub use tendermint_proto::serializers::*;
 pub mod apphash;
 pub mod hash;
 pub mod option_hash;
-pub mod time;