+++ CHANGELOG.md
@@ -15,6 +15,8 @@
   deemed valid instead of invalid ([#650])
 - `[light-client]` Revert a change introduced in [#652] that would enable DoS attacks,
   where full nodes could spam the light client with massive commits (eg. 10k validators).
+- `[tendermint]` (Since RC1) Time serialization fix (part of [#665])
+- `[tendermint-proto]` (Since RC1) Timestamp serialization fix (part of [#665])
 
 ### FEATURES:
 
@@ -25,6 +27,7 @@
 [#652]: https://github.com/informalsystems/tendermint-rs/pulls/652
 [#639]: https://github.com/informalsystems/tendermint-rs/pull/639
 [#660]: https://github.com/informalsystems/tendermint-rs/issues/660
+[#665]: https://github.com/informalsystems/tendermint-rs/issues/665
 
 ## v0.17.0-rc1